+++ lti_consumer/models.py
@@ -34,6 +34,7 @@ from lti_consumer.utils import (
     choose_lti_1p3_redirect_uris,
     model_to_dict,
     EXTERNAL_ID_REGEX,
+    external_multiple_launch_urls_enabled,
 )
 
 log = logging.getLogger(__name__)
@@ -586,10 +587,17 @@ class LtiConfiguration(models.Model):
                 tool_keyset_url=self.lti_1p3_tool_keyset_url,
             )
         elif self.config_store == self.CONFIG_EXTERNAL:
+            lti_launch_url = self.external_config.get('lti_1p3_launch_url')
+
+            if external_multiple_launch_urls_enabled(self.location.course_key):
+                block = compat.load_enough_xblock(self.location)
+
+                lti_launch_url = block.lti_1p3_launch_url or lti_launch_url
+
             consumer = consumer_class(
                 iss=get_lti_api_base(),
                 lti_oidc_url=self.external_config.get('lti_1p3_oidc_url'),
-                lti_launch_url=self.external_config.get('lti_1p3_launch_url'),
+                lti_launch_url=lti_launch_url,
                 client_id=self.external_config.get('lti_1p3_client_id'),
                 # Deployment ID hardcoded to 1 since
                 # we're not using multi-tenancy.
@@ -695,7 +703,7 @@ class LtiAgsLineItem(models.Model):
     )
 
     # Tool resource identifier, not used by the LMS.
-    resource_id = models.CharField(max_length=100, blank=True)
+    resource_id = models.CharField(max_length=255, blank=True)
 
     # LMS Resource link
     # Must be the same as the one sent in the tool's LTI launch.
@@ -824,6 +832,8 @@ class LtiDlContentItem(models.Model):
     LTI-DL Specification: https://www.imsglobal.org/spec/lti-dl/v2p0
     Content items are resources selected by instructor that should
     be displayed to students.
+
+    .. no_pii:
     """
     # LTI Configuration link
     # This ties the LineItem to each tool configuration
@@ -920,4 +930,5 @@ class CourseAllowPIISharingInLTIFlag(ConfigurationModel):
     class Meta:
         # This model was moved from edx-platform, with intention of retaining existing data.
         # This is referencing the original table name.
+        app_label = "lti_consumer"
         db_table = "xblock_config_courseeditltifieldsenabledflag"