+++ lti_consumer/tests/unit/plugin/test_views.py
@@ -5,13 +5,13 @@ import json
 from unittest.mock import patch, Mock
 
 import ddt
-
+import jwt
 from django.test.testcases import TestCase
 from django.urls import reverse
 from edx_django_utils.cache import TieredCache, get_cache_key
+from jwt.api_jwk import PyJWK
 
 from Cryptodome.PublicKey import RSA
-from jwkest.jwk import RSAKey
 from opaque_keys.edx.keys import UsageKey
 from lti_consumer.data import Lti1p3LaunchData, Lti1p3ProctoringLaunchData
 from lti_consumer.models import LtiConfiguration, LtiDlContentItem
@@ -179,6 +179,13 @@ class TestLti1p3LaunchGateEndpoint(TestCase):
         self.compat.get_course_by_id.return_value = course
         self.compat.get_user_role.return_value = "student"
         self.compat.get_external_id_for_user.return_value = "12345"
+
+        block_compat_patcher = patch("lti_consumer.lti_xblock.compat")
+        self.addCleanup(block_compat_patcher.stop)
+        block_compat = block_compat_patcher.start()
+        block_compat.get_course_by_id.return_value = course
+        block_compat.get_user_role.return_value = "student"
+        block_compat.get_external_id_for_user.return_value = "12345"
 
         model_compat_patcher = patch("lti_consumer.models.compat")
         self.addCleanup(model_compat_patcher.stop)
@@ -674,8 +681,14 @@ class TestLti1p3AccessTokenEndpoint(TestCase):
         )
         self.addCleanup(get_lti_consumer_patcher.stop)
         self._mock_xblock_handler = get_lti_consumer_patcher.start()
-        # Generate RSA
-        self.key = RSAKey(key=RSA.generate(2048), kid="1")
+        # Generate RSA and save exports
+        rsa_key = RSA.generate(2048)
+        algo_obj = jwt.get_algorithm_by_name('RS256')
+        private_key = algo_obj.prepare_key(rsa_key.export_key())
+        private_jwk = json.loads(algo_obj.to_jwk(private_key))
+        private_jwk['kid'] = 1
+        self.key = PyJWK.from_dict(private_jwk)
+        self.public_key = rsa_key.public_key().export_key('PEM')
 
     def get_body(self, token, **overrides):
         """
@@ -697,9 +710,13 @@ class TestLti1p3AccessTokenEndpoint(TestCase):
         self.mock_client.access_token.return_value = token
 
         body = self.get_body(create_jwt(self.key, {}))
-        response = self.client.post(self.url, data=body)
+        response = self.client.post(self.url, data=json.dumps(body), content_type='application/json')
+        self.mock_client.access_token.assert_called_once()
+        called_args = self.mock_client.access_token.call_args[0]
+        actual_arg = called_args[0]
+        actual_dict = json.loads(next(iter(actual_arg.keys())))
 
-        self.mock_client.access_token.called_once_with(body)
+        self.assertEqual(actual_dict, body)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), token)
 
@@ -715,9 +732,14 @@ class TestLti1p3AccessTokenEndpoint(TestCase):
             args=[str(self.config.location)]
         )
         body = self.get_body(create_jwt(self.key, {}))
-        response = self.client.post(url, data=body)
+        response = self.client.post(url, data=json.dumps(body), content_type='application/json')
 
-        self.mock_client.access_token.called_once_with(body)
+        self.mock_client.access_token.assert_called_once()
+        called_args = self.mock_client.access_token.call_args[0]
+        actual_arg = called_args[0]
+        actual_dict = json.loads(next(iter(actual_arg.keys())))
+
+        self.assertEqual(actual_dict, body)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), token)
 
@@ -748,7 +770,8 @@ class TestLti1p3AccessTokenEndpoint(TestCase):
         body = self.get_body(create_jwt(self.key, {}))
         response = self.client.post(
             reverse('lti_consumer:lti_consumer.access_token_via_external_id', args=['x', 'x']),
-            data=body,
+            data=json.dumps(body),
+            content_type='application/json'
         )
 
         get_external_config_from_filter.assert_called_once_with({}, 'x:x')
@@ -765,7 +788,12 @@ class TestLti1p3AccessTokenEndpoint(TestCase):
             tool_key=external_config['lti_1p3_tool_public_key'],
             tool_keyset_url=external_config['lti_1p3_tool_keyset_url'],
         )
-        lti_consumer().access_token.called_once_with(body)
+        lti_consumer().access_token.assert_called_once()
+        called_args = lti_consumer().access_token.call_args[0]
+        actual_arg = called_args[0]
+        actual_dict = json.loads(next(iter(actual_arg.keys())))
+
+        self.assertEqual(actual_dict, body)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), token)